refactor: extract iOS installment parsing into standalone class
Create InstallmentConfigurationParser for iOS following existing patterns:
- Takes NSDictionary in initializer
- Uses computed property for installmentConfiguration
- Encapsulates installment parsing logic
- CardConfigurationParser delegates to InstallmentConfigurationParser
- Improves code modularity and consistency across platforms
